Create and run Webpack using a basic installation.
Create a folder named task_0.
Install webpack and webpack-cli as developer dependencies within the folder using npm.
Install jQuery as a regular dependency using npm.
Make sure that webpack and webpack-cli are listed under the devDependencies key along with jQuery being listed under the dependencies key within the package.json
Create a src directory with a index.js in it.
The file should import jquery and add three different paragraphs to the page body: (refer to this link under the Babel section for importing jQuery)