Create a specific configuration for Webpack.
Using the folder named task_2.
Reuse the code from task_1.
Modify the webpack config to support adding CSS to the bundle.
Modify the webpack config to support adding images to the CSS.
Create a folder named css
In a file named main.css, change the position of the counter text to be displayed on the right of the button and in bold.
Add an element at the top of document with the id #logo which has a width of 200px and height of 200px.
Set the background of th element with the image in task_2/assets/holberton-logo.jpg (you can download this link).
Set the width and height of the logo background image to 200px by 200px.
Make sure to configure Webpack to optimize images.
When running Webpack, your javascript and html files should be generated in a public folder.
The javascript file generated by Webpack should be named bundle.js.
Set Webpack config mode to production.
Opening your main file should not generate any error in the console.
Your html code should only import one Javascript script (the one generated by webpack).
When running Webpack, you should not see the warning WARNING in asset size limit: The following asset(s) exceed the recommended size limit