Let’s create an animation that we can display when the user hovers on the Notifications menu or when there is a new notification. In task_4/dashboard/Notifications/Notifications.js:
Create one object containing the CSS frames to make the opacity change from 0.5 to 1
Create one object containing the CSS frames to make the element bounce. You can play with translateY and alternate from 0px to -5px and 5px
Modify the styling for the menu item to:

Float on the right of the screen over every element
The background color should be #fff8f8
Show the pointer cursor when hovering the element
On hover, animate the element with the two new animations. The duration for the opacity change should be 1s, and the duration for the bouncing effect should be 0.5s. The animation should repeat 3 times only
When the list of notifications is visible, hide the menu item
Requirements:

When the notifications panel is hovered or opened, the UI should look like the image below: